Organizations for git-port and AC tests.
The Aho Corasick tests are running from Eclipse but somehow not from the
command line with Maven. That is, the tests will run from the command
line but the results will be wrong. This is not handeled for the moment
as the AC plugin would have to be adapted to newer Neo4j APIs anyway and
is currently unused.