Did quite some restructuring.
There are issues with missing taxonomical relationships for BioPortal
ontologies. This might be caused by the fact that at some point, when
aggregates were made directly importable, the tracking of already
existing concept nodes was done at the wrong place, in the
insertFacetTerm method. This is still subject to tests.

Removed the nodeBySrcId map in favor of a nodyByCoordinates map. This
map is now filled at only two places. The first and major place is the
second insertFacetTerms method (the "first" is the public method that
then calls the "second", private method). The second place is the
createRelationships method and then only for HOLLOW parents.
All tests working.