Added log messages and fixed aggregates.
Aggregate insertion was formerly omitted from the general strategy of
concept insertion because aggregate are not required to have
coordinates. However, this doesn't make sense, now we just omit those
concepts that don't have coordinates but include aggregates in general.
This is necessary for concepts that have aggregates as parents.