Removed the Aho-Corasick plugin because all tests failed in 3.2.3.
After the Neo4j 3.2.3 update and the correction of all syntax errors in
the Aho Corasick project, no tests would run correctly any more. Since
we are currently not using the plugin, it was removed. If it is requried
again in the future, it must be fetched from the last revision and be
repaired. It is completely unknown why exactly the algorithm would
behave differently. It is also possible that only the correction of the
syntax errors was wrong, i.e. some corrections did change the semantics
of the code.