Using Jackson.
Since we now allow the concepts in ImportConcepts to be a Java 8 stream,
gson did no longer work since it could not (or I didn't find how to get
it to) properly handle streams. Thus, the switch to Jackson which is
able to do it using a specific java 8 module.