Allow the specification of aggregate copy properties for equal name agg creation and make them more efficient.
More efficiency by accumulating the new properties in a in-memory map instead of writing to the node all the time. I am actually not sure that this helps since transactions are made in memory anyway. But it seemed to improve things a bit. The actual issue was a missing index, however. Since the copy property process accessed the id property of AGGREGATE nodes, the corresponding index is tremendously helpful.